Prep for 0.2.0 release (deployed from strike-development staging)

+++ aura/strike_pill/strike_pillController.js
@@ -1,10 +1,8 @@
 ({
-	onInit: function(component, event, helper) {
-		
-		
-	},
-	handleClick: function(component,event,helper){
-		helper.notifyParent(component);
-		helper.destroyComponent(component);
+	destroy: function(component, event, helper) {
+		if (component.get('v.isDestroyable')) {
+			helper.notifyParent(component);
+			helper.destroyComponent(component);
+		}
 	}
 })
